+++ extras/gluonts/models/tinytimemixer/__init__.py
@@ -1,4 +1,4 @@
-from tsfm_public.models.tinytimemixer.gluonts.ttm_gluonts_predictor import (
+from .ttm_gluonts_predictor import (
     TTM_AVAILABLE_CONTEXTS,
     TTM_MAX_FORECAST_HORIZON,
     TTMGluonTSPredictor,
